clean class methods
