+++ src/modules/add-remove.js
@@ -1,51 +1,10 @@
 import {
-  deleteAllTasks, updateInputState, makeInputDefault,
+  updateInputState, makeInputDefault,
 } from './interactive-page.js';
 
-const taskContainer = document.querySelector('.task-container');
-const enterBtn = document.querySelector('.enter');
-export const clearAllBtn = document.getElementById('clear-btn');
-const refresh = document.querySelector('.refresh');
 class AddRemoveTask {
   constructor() {
     this.store = JSON.parse(localStorage.getItem('ToDoList')) || [];
-  }
-
-  resetIndex() {
-    let initialIndex = 1;
-    this.store.forEach((task) => {
-      task.index = initialIndex;
-      initialIndex += 1;
-    });
-    localStorage.setItem('ToDoList', JSON.stringify(this.store));
-  }
-
-  deleteTask(index) {
-    this.store = this.store.filter((task) => task.index !== index);
-    this.resetIndex();
-    this.newTask();
-    localStorage.setItem('ToDoList', JSON.stringify(this.store));
-  }
-
-  editTask(index, value) {
-    this.store[index].description = value;
-    localStorage.setItem('ToDoList', JSON.stringify(this.store));
-  }
-
-  markItemAsComplete(index) {
-    if (this.store[index].completed) {
-      this.store[index].completed = false;
-    } else {
-      this.store[index].completed = true;
-    }
-
-    localStorage.setItem('ToDoList', JSON.stringify(this.store));
-  }
-
-  clearCompletedTasks() {
-    this.store = this.store.filter((task) => task.completed === false);
-    this.resetIndex();
-    this.newTask();
   }
 
   newTask() {
@@ -77,9 +36,74 @@ class AddRemoveTask {
       }
     });
 
+    this.pageEvents();
+    localStorage.setItem('ToDoList', JSON.stringify(this.store));
+  }
+
+  resetIndex() {
+    let initialIndex = 1;
+    this.store.forEach((task) => {
+      task.index = initialIndex;
+      initialIndex += 1;
+    });
+    localStorage.setItem('ToDoList', JSON.stringify(this.store));
+  }
+  
+  addNewTask(task) {
+    const currTask = {
+      index: this.store.length < 1 ? 1 : this.store.length + 1,
+      description: task,
+      completed: false,
+    };
+    this.store.push(currTask);
+    this.newTask();
+    localStorage.setItem('ToDoList', JSON.stringify(this.store));
+  }
+
+  deleteTask(index) {
+    this.store = this.store.filter((task) => task.index !== index);
+    this.resetIndex();
+    this.newTask();
+    localStorage.setItem('ToDoList', JSON.stringify(this.store));
+  }
+
+  editTask(index, value) {
+    this.store[index].description = value;
+    localStorage.setItem('ToDoList', JSON.stringify(this.store));
+  }
+
+  markItemAsComplete(index) {
+    if (this.store[index].completed) {
+      this.store[index].completed = false;
+    } else {
+      this.store[index].completed = true;
+    }
+
+    localStorage.setItem('ToDoList', JSON.stringify(this.store));
+  }
+
+  clearCompletedTasks() {
+    this.store = this.store.filter((task) => task.completed === false);
+    this.resetIndex();
+    this.newTask();
+  }
+
+  clearAllTasks() {
+    this.store = []
+    localStorage.setItem('ToDoList', JSON.stringify(this.store));
+    this.newTask()
+  }
+
+  localStorageToWebpage() {
+    const store = JSON.parse(localStorage.getItem('ToDoList'));
+    store.forEach((task) => {
+      this.newTask(task.description);
+    });
+  }
+
+  pageEvents() { 
     // mark item as completed
     const checkBox = document.querySelectorAll('.check');
-
     checkBox.forEach((box) => {
       box.addEventListener('change', () => {
         if (box.checked) {
@@ -89,15 +113,6 @@ class AddRemoveTask {
           box.nextElementSibling.classList.remove('line');
           this.markItemAsComplete(Number(box.nextElementSibling.id) - 1, this.store);
         }
-      });
-    });
-
-    // delete an item
-    const deleteBtn = document.querySelectorAll('.delete');
-    deleteBtn.forEach((btn) => {
-      btn.addEventListener('click', () => {
-        taskContainer.innerHTML = '';
-        this.deleteTask(Number(btn.parentElement.id));
       });
     });
 
@@ -120,60 +135,13 @@ class AddRemoveTask {
       });
     });
 
-    localStorage.setItem('ToDoList', JSON.stringify(this.store));
-  }
-
-  localStorageToWebpage() {
-    const store = JSON.parse(localStorage.getItem('ToDoList'));
-    store.forEach((task) => {
-      this.newTask(task.description);
+    // delete an item
+    const deleteBtn = document.querySelectorAll('.delete');
+    deleteBtn.forEach((btn) => {
+      btn.addEventListener('click', () => {
+        this.deleteTask(Number(btn.parentElement.id));
+      });
     });
-  }
-
-  addNewTask(task) {
-    const currTask = {
-      index: this.store.length < 1 ? 1 : this.store.length + 1,
-      description: task,
-      completed: false,
-    };
-    this.store.push(currTask);
-    this.newTask();
-    localStorage.setItem('ToDoList', JSON.stringify(this.store));
-  }
-
-  submitNewTaskEntry() {
-    const addNewTaskInput = document.getElementById('add-new-task');
-    const errormsg = document.getElementById('error-msg');
-
-    enterBtn.addEventListener('click', () => {
-      if (addNewTaskInput.value === '') {
-        errormsg.style.visibility = 'visible';
-        setTimeout(() => {
-          errormsg.style.visibility = 'hidden';
-        }, 2000);
-      } else {
-        this.addNewTask(addNewTaskInput.value);
-        addNewTaskInput.value = '';
-      }
-    });
-
-    addNewTaskInput.addEventListener('keydown', (e) => {
-      if (e.code === 'Enter') {
-        if (addNewTaskInput.value === '') {
-          errormsg.style.visibility = 'visible';
-          setTimeout(() => {
-            errormsg.style.visibility = 'hidden';
-          }, 2000);
-        } else {
-          this.addNewTask(addNewTaskInput.value);
-          addNewTaskInput.value = '';
-        }
-      }
-    });
-
-    // delete all tasks
-    deleteAllTasks(refresh, taskContainer, this.store);
-    localStorage.setItem('ToDoList', JSON.stringify(this.store));
   }
 }
